# src/main/java/com/lockbeck/entities/subject/SubjectController.java
package com.lockbeck.entities.subject;

import lombok.RequiredArgsConstructor;
import lombok.extern.log4j.Log4j2;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping("/subject")
@RequiredArgsConstructor
@Validated
@Log4j2
public class SubjectController {
    private final SubjectService service;
    @PostMapping("/create")
    public ResponseEntity<?> create(@RequestBody SubjectCreateRequest request){
        return ResponseEntity.ok(service.create(request));
    }

    @GetMapping("/list")
    public ResponseEntity<?> list(){
        return ResponseEntity.ok(service.list());
    }
}